data-pipeline/requirements.txt
pandas>=2.1.0
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
boto3>=1.34.0
sqlalchemy>=2.0.0
tqdm>=4.65.0